+++ knf/validators/network/validators_test.go
@@ -8,7 +8,7 @@ package network
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"io/ioutil"
+	"os"
 	"testing"
 
 	"github.com/essentialkaos/ek/v12/knf"
@@ -174,7 +174,7 @@ func (s *ValidatorSuite) TestURLValidator(c *C) {
 func createConfig(c *C, data string) string {
 	configPath := c.MkDir() + "/config.knf"
 
-	err := ioutil.WriteFile(configPath, []byte(data), 0644)
+	err := os.WriteFile(configPath, []byte(data), 0644)
 
 	if err != nil {
 		c.Fatal(err.Error())