+++ knf/validators/regexp/validators_test.go
@@ -8,7 +8,7 @@ package regexp
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"io/ioutil"
+	"os"
 	"testing"
 
 	"github.com/essentialkaos/ek/v12/knf"
@@ -72,7 +72,7 @@ func (s *ValidatorSuite) TestRegexpValidator(c *check.C) {
 func createConfig(c *check.C, data string) string {
 	configPath := c.MkDir() + "/config.knf"
 
-	err := ioutil.WriteFile(configPath, []byte(data), 0644)
+	err := os.WriteFile(configPath, []byte(data), 0644)
 
 	if err != nil {
 		c.Fatal(err.Error())