+++ system/info_linux_test.go
@@ -9,7 +9,6 @@ package system
 
 import (
 	"bufio"
-	"io/ioutil"
 	"os"
 	"strconv"
 	"strings"
@@ -703,7 +702,7 @@ func (s *SystemSuite) CreateTestFile(c *C, data string) string {
 	tmpDir := c.MkDir()
 	tmpFile := tmpDir + "/test.file"
 
-	if ioutil.WriteFile(tmpFile, []byte(data), 0644) != nil {
+	if os.WriteFile(tmpFile, []byte(data), 0644) != nil {
 		c.Fatal("Can't create temporary file")
 	}