+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 12.60.0
+
+* `[system/container]` Added container sub-package with methods for checking container engine info
+* `[system]` Added container engine info to `SystemInfo`
+* `[fmtutil/table]` Improved separator rendering
+* Code refactoring
+
 ### 12.59.0
 
 * `[fmtutil/table]` Improved separator rendering